+++ src/components/KeyShortchut.tsx
@@ -2,7 +2,7 @@ import type { ReactNode } from 'react';
 
 export const KeyShortcut = ({ children }: { children: ReactNode }) => {
   return (
-    <kbd className="px-2 text-xs text-white bg-indigo-600 border-slate-700 rounded-lg">
+    <kbd className="rounded-lg border-slate-700 bg-indigo-600 px-2 text-xs text-white">
       {children}
     </kbd>
   );