+++ src/components/StreamChatSelectInput.tsx
@@ -12,9 +12,9 @@ export const StreamChatSelectInput = () => {
   const setSelectedStream = useStore((state) => state.setSelectedStream);
 
   return (
-    <div className="w-80 my-4 lg:my-0 relative">
+    <div className="relative my-4 w-80 lg:my-0">
       <Listbox value={selectedStream} onChange={setSelectedStream}>
-        <Listbox.Button className="relative w-full cursor-default rounded-lg bg-slate-700 py-2 pl-3 pr-10 hover:cursor-pointer text-left shadow-md focus:outline-none focus-visible:border-indigo-500 focus-visible:ring-2 focus-visible:ring-white focus-visible:ring-opacity-75 focus-visible:ring-offset-2 focus-visible:ring-offset-orange-300 sm:text-sm">
+        <Listbox.Button className="relative w-full cursor-default rounded-lg bg-slate-700 py-2 pl-3 pr-10 text-left shadow-md hover:cursor-pointer focus:outline-none focus-visible:border-indigo-500 focus-visible:ring-2 focus-visible:ring-white focus-visible:ring-opacity-75 focus-visible:ring-offset-2 focus-visible:ring-offset-orange-300 sm:text-sm">
           <span className="block truncate">{selectedStream.channel}</span>
           <span className="pointer-events-none absolute inset-y-0 right-0 flex items-center pr-2">
             <ChevronsUpDown
@@ -44,7 +44,7 @@ export const StreamChatSelectInput = () => {
                 {({ selected }) => (
                   <>
                     <span
-                      className={clsx('block text-white truncate', {
+                      className={clsx('block truncate text-white', {
                         'font-medium': selected,
                         'font-normal': !selected,
                       })}